feat: auto switch to mainnet on shell (#1135)
